AI/pyproject.toml
[build-system]
requires = ["flit_core>=3.10"]
build-backend = "flit_core.buildapi"


[project]
name = "travel_lotara"
description = "A multi-agent travel assistant powered by Google ADK & Opik"
readme = "README.md"
requires-python = ">=3.10,<3.15"
# license = { file = "LICENSE" }
authors = [{ name = "Lotara Organization", email = "lotara@lotara.com" }]
dynamic = ["version"]

classifiers = [
  "Typing :: Typed",
  "Intended Audience :: Developers",
  "Intended Audience :: Science/Research",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Operating System :: OS Independent",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "License :: OSI Approved :: Apache Software License",
]

dependencies = [
  "PyYAML>=6.0.3",
  "pydantic>=2.6.0",
  "opik>=1.0.0",
  "google-adk>=0.1.0",
  "supabase>=2.0.0",
  "openinference-instrumentation-google-adk>=0.1.0; python_version >= '3.11' and python_version < '3.14'",
  "openinference-instrumentation>=0.1.34",
]

[tool.uv]
link-mode = "copy"

[project.urls]
homepage = "https://google.github.io/lotara-docs/"
repository = "https://github.com/nminhkhoi0818/lotara"
changelog = "https://github.com/nminhkhoi0818/lotara/blob/main/CHANGELOG.md"
documentation = "https://google.github.io/lotara-docs/"


[project.scripts]
lotara_ai = "lotara.ai.main:main"


# ------------------------------------------------------------------
# OPTIONAL DEPENDENCIES (EXTRAS)
# ------------------------------------------------------------------

[project.optional-dependencies]

# Development & quality
dev = [
  "flit>=3.10.0",
  "isort>=6.0.0",
  "mypy>=1.15.0",
  "pyink>=24.10.0",
  "pylint>=2.6.0",
]

# LLM providers & adapters
llm = [
  "anthropic>=0.43.0",
  "litellm>=1.75.5,<2.0.0",
]

# Agent frameworks / orchestration
agents = [
  "a2a-sdk>=0.3.4,<0.4.0",
  "langgraph>=0.2.60,<0.4.8",
  "crewai[tools]; python_version>='3.11' and python_version<'3.12'",
]

# Retrieval & indexing
retrieval = [
  "llama-index-readers-file>=0.4.0",
  "llama-index-embeddings-google-genai>=0.3.0",
]

# Web / parsing tools
web = [
  "beautifulsoup4>=3.2.2",
  "lxml>=5.3.0",
]

# Evaluation & benchmarking
eval = [
  "pandas>=2.2.3",
  "rouge-score>=0.1.2",
  "tabulate>=0.9.0",
  "tqdm>=4.65.0",
  "scipy<1.16; python_version<'3.11'",
]

# Testing
test = [
  "pytest>=9.0.0,<10.0.0",
  "pytest-asyncio>=0.25.0",
  "pytest-mock>=3.14.0",
  "pytest-xdist>=3.6.1",
  "python-multipart>=0.0.9",
  "openai>=1.100.2",
]

# Documentation
docs = [
  "autodoc_pydantic",
  "furo",
  "myst-parser",
  "sphinx",
  "sphinx-autodoc-typehints",
  "sphinx-rtd-theme",
]

# Tooling integrations
toolbox = [
  "toolbox-adk>=0.5.7,<0.6.0",
]

otel-gcp = [
  "opentelemetry-instrumentation-google-genai>=0.3b0,<1.0.0",
]


# ------------------------------------------------------------------
# TOOL CONFIGURATION
# ------------------------------------------------------------------

[tool.pyink]
line-length = 80
unstable = true
pyink-indentation = 2
pyink-use-majority-quotes = true
pyink-annotation-pragmas = [
  "noqa",
  "pylint:",
  "type: ignore",
  "pytype:",
  "mypy:",
  "pyright:",
  "pyre-",
]


[tool.isort]
profile = "lotara"
single_line_exclusions = []
line_length = 200
known_first_party = ["lotara", "google"]


[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_default_fixture_loop_scope = "function"
asyncio_mode = "auto"


[tool.mypy]
python_version = "3.10"
exclude = "tests/"
plugins = ["pydantic.mypy"]
disable_error_code = [
  "import-not-found",
  "import-untyped",
  "unused-ignore",
]
follow_imports = "skip"
